moved configuration directly to Connection options
